Abstract class for "tagIDs" filter for CPTs
