+++ layers/CMSSchema/packages/categories/src/SchemaHooks/AbstractAddCategoryFilterInputObjectTypeHookSet.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PoPCMSSchema\PostTags\SchemaHooks;
+namespace PoPCMSSchema\Categories\SchemaHooks;
 
 use PoP\ComponentModel\FilterInputs\FilterInputInterface;
 use PoP\ComponentModel\Schema\SchemaTypeModifiers;
@@ -10,19 +10,14 @@ use PoP\ComponentModel\TypeResolvers\InputObjectType\HookNames;
 use PoP\ComponentModel\TypeResolvers\InputObjectType\InputObjectTypeResolverInterface;
 use PoP\ComponentModel\TypeResolvers\InputTypeResolverInterface;
 use PoP\ComponentModel\TypeResolvers\ScalarType\IDScalarTypeResolver;
-use PoP\ComponentModel\TypeResolvers\ScalarType\StringScalarTypeResolver;
 use PoP\Root\App;
 use PoP\Root\Hooks\AbstractHookSet;
-use PoPCMSSchema\Posts\TypeResolvers\InputObjectType\PostsFilterInputObjectTypeResolverInterface;
-use PoPCMSSchema\Tags\FilterInputs\TagIDsFilterInput;
-use PoPCMSSchema\Tags\FilterInputs\TagSlugsFilterInput;
+use PoPCMSSchema\Categories\FilterInputs\CategoryIDsFilterInput;
 
-class InputObjectTypeHookSet extends AbstractHookSet
+abstract class AbstractAddCategoryFilterInputObjectTypeHookSet extends AbstractHookSet
 {
     private ?IDScalarTypeResolver $idScalarTypeResolver = null;
-    private ?StringScalarTypeResolver $stringScalarTypeResolver = null;
-    private ?TagSlugsFilterInput $tagSlugsFilterInput = null;
-    private ?TagIDsFilterInput $tagIDsFilterInput = null;
+    private ?CategoryIDsFilterInput $categoryIDsFilterInput = null;
 
     final public function setIDScalarTypeResolver(IDScalarTypeResolver $idScalarTypeResolver): void
     {
@@ -33,32 +28,14 @@ class InputObjectTypeHookSet extends AbstractHookSet
         /** @var IDScalarTypeResolver */
         return $this->idScalarTypeResolver ??= $this->instanceManager->getInstance(IDScalarTypeResolver::class);
     }
-    final public function setStringScalarTypeResolver(StringScalarTypeResolver $stringScalarTypeResolver): void
+    final public function setCategoryIDsFilterInput(CategoryIDsFilterInput $categoryIDsFilterInput): void
     {
-        $this->stringScalarTypeResolver = $stringScalarTypeResolver;
+        $this->categoryIDsFilterInput = $categoryIDsFilterInput;
     }
-    final protected function getStringScalarTypeResolver(): StringScalarTypeResolver
+    final protected function getCategoryIDsFilterInput(): CategoryIDsFilterInput
     {
-        /** @var StringScalarTypeResolver */
-        return $this->stringScalarTypeResolver ??= $this->instanceManager->getInstance(StringScalarTypeResolver::class);
-    }
-    final public function setTagSlugsFilterInput(TagSlugsFilterInput $tagSlugsFilterInput): void
-    {
-        $this->tagSlugsFilterInput = $tagSlugsFilterInput;
-    }
-    final protected function getTagSlugsFilterInput(): TagSlugsFilterInput
-    {
-        /** @var TagSlugsFilterInput */
-        return $this->tagSlugsFilterInput ??= $this->instanceManager->getInstance(TagSlugsFilterInput::class);
-    }
-    final public function setTagIDsFilterInput(TagIDsFilterInput $tagIDsFilterInput): void
-    {
-        $this->tagIDsFilterInput = $tagIDsFilterInput;
-    }
-    final protected function getTagIDsFilterInput(): TagIDsFilterInput
-    {
-        /** @var TagIDsFilterInput */
-        return $this->tagIDsFilterInput ??= $this->instanceManager->getInstance(TagIDsFilterInput::class);
+        /** @var CategoryIDsFilterInput */
+        return $this->categoryIDsFilterInput ??= $this->instanceManager->getInstance(CategoryIDsFilterInput::class);
     }
 
     protected function init(): void
@@ -97,29 +74,29 @@ class InputObjectTypeHookSet extends AbstractHookSet
         array $inputFieldNameTypeResolvers,
         InputObjectTypeResolverInterface $inputObjectTypeResolver,
     ): array {
-        if (!($inputObjectTypeResolver instanceof PostsFilterInputObjectTypeResolverInterface)) {
+        if (!(is_a($inputObjectTypeResolver, $this->getInputObjectTypeResolverClass(), true))) {
             return $inputFieldNameTypeResolvers;
         }
         return array_merge(
             $inputFieldNameTypeResolvers,
             [
-                'tagIDs' => $this->getIDScalarTypeResolver(),
-                'tagSlugs' => $this->getStringScalarTypeResolver(),
+                'categoryIDs' => $this->getIDScalarTypeResolver(),
             ]
         );
     }
+
+    abstract protected function getInputObjectTypeResolverClass(): string;
 
     public function getInputFieldDescription(
         ?string $inputFieldDescription,
         InputObjectTypeResolverInterface $inputObjectTypeResolver,
         string $inputFieldName
     ): ?string {
-        if (!($inputObjectTypeResolver instanceof PostsFilterInputObjectTypeResolverInterface)) {
+        if (!(is_a($inputObjectTypeResolver, $this->getInputObjectTypeResolverClass(), true))) {
             return $inputFieldDescription;
         }
         return match ($inputFieldName) {
-            'tagIDs' => $this->__('Get results from the tags with given IDs', 'pop-users'),
-            'tagSlugs' => $this->__('Get results from the tags with given slug', 'pop-users'),
+            'categoryIDs' => $this->__('Get results from the categories with given IDs', 'pop-users'),
             default => $inputFieldDescription,
         };
     }
@@ -129,12 +106,11 @@ class InputObjectTypeHookSet extends AbstractHookSet
         InputObjectTypeResolverInterface $inputObjectTypeResolver,
         string $inputFieldName
     ): int {
-        if (!($inputObjectTypeResolver instanceof PostsFilterInputObjectTypeResolverInterface)) {
+        if (!(is_a($inputObjectTypeResolver, $this->getInputObjectTypeResolverClass(), true))) {
             return $inputFieldTypeModifiers;
         }
         return match ($inputFieldName) {
-            'tagIDs',
-            'tagSlugs'
+            'categoryIDs'
                 => SchemaTypeModifiers::IS_ARRAY | SchemaTypeModifiers::IS_NON_NULLABLE_ITEMS_IN_ARRAY,
             default
                 => $inputFieldTypeModifiers,
@@ -146,12 +122,11 @@ class InputObjectTypeHookSet extends AbstractHookSet
         InputObjectTypeResolverInterface $inputObjectTypeResolver,
         string $inputFieldName,
     ): ?FilterInputInterface {
-        if (!($inputObjectTypeResolver instanceof PostsFilterInputObjectTypeResolverInterface)) {
+        if (!(is_a($inputObjectTypeResolver, $this->getInputObjectTypeResolverClass(), true))) {
             return $inputFieldFilterInput;
         }
         return match ($inputFieldName) {
-            'tagIDs' => $this->getTagIDsFilterInput(),
-            'tagSlugs' => $this->getTagSlugsFilterInput(),
+            'categoryIDs' => $this->getCategoryIDsFilterInput(),
             default => $inputFieldFilterInput,
         };
     }
